// light-modules/spa-next-app/brochure_next-app/types/commonTypes.ts
export interface HeaderType {
  brandText: string;
  salesTeamText: string;
  salesTeaNumber: string;
  salesTeamOpeningTime: string;
  CustomerServiceText: string;
  CustomerServiceNumber: string;
  CustomerServiceOpeningTime: string;
  beCoveredText: string;
  getAQuoteText: string;
  loginText: string;
  openingTimesSalesTeam: string;
  openingTimesCustomerService: string;
  openingTimesCustomerServiceHeader: string;
  openingTimesSalesTeamHeader: string;
  travelInsurance: string;
  otherProducts: string;
  whyStaysure: string;
  claims: string;
  blog: string;
  help: string;
  travelInsuranceMenu1: string;
  travelInsuranceMenu2: string;
  travelInsuranceMenu3: string;
  travelInsuranceMenu4: string;
  travelInsuranceMenu5: string;
  travelInsuranceMenu6: string;
  travelInsuranceMenu7: string;
  travelInsuranceMenu8: string;
  travelInsuranceMenu9: string;
  travelInsuranceMenu10: string;
  travelInsuranceMenu11: string;
  travelInsuranceMenu12: string;
  helpMenu1: string;
}

export interface FooterType {
  blockOneHeadline: string;
  blockOneLink1: string;
  blockOneLink2: string;
  blockOneLink3: string;
  blockOneLink4: string;
  blockOneLink5: string;
  blockOneLink6: string;
  blockOneLink7: string;
  blockTwoHeadline: string;
  blockTwoLink1: string;
  blockTwoLink2: string;
  blockTwoLink3: string;
  blockTwoLink4: string;
  blockTwoLink5: string;
  blockTwoLink6: string;
  blockTwoLink7: string;
  blockTwoLink8: string;
  blockThreeHeadline: string;
  blockThreeCopyRight: string;
  footerContent: string | TrustedHTML;
}